backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
pydantic-core
numpy
scipy
sentence-transformers
PyPDF2
python-docx
pdf2image
pillow
pytesseract
weasyprint
jinja2
spacy
pytest
redis
spacy